fix: refactor photo_processing structure and update all references
