+++ soil_analysis/domain/service/photo_processing.py
@@ -1,13 +1,12 @@
 from haversine import haversine, Unit
 
 from lib.geo.valueobject.coord import XarvioCoord
-from soil_analysis.domain.valueobject.photo import AndroidPhoto
-from soil_analysis.domain.valueobject.photo_land_association import PhotoLandAssociation
-from soil_analysis.domain.valueobject.photo_spot import PhotoSpot
+from soil_analysis.domain.valueobject.photo_processing.photo import AndroidPhoto
+from soil_analysis.domain.valueobject.photo_processing.photo_land_association import PhotoLandAssociation
+from soil_analysis.domain.valueobject.photo_processing.photo_spot import PhotoSpot
 from soil_analysis.models import Land
 
 
-# TODO: ファイル名を photo_processing.py に変更 Issue284
 class PhotoProcessingService:
     def process_photos(
         self, photo_path_list: list[str], land_list: list[Land]
