+++ soil_analysis/domain/valueobject/photo_processing/photo.py
@@ -5,7 +5,7 @@ import re
 import exifread
 
 from lib.geo.valueobject.coord import XarvioCoord
-from soil_analysis.domain.valueobject.photo_spot import PhotoSpot
+from soil_analysis.domain.valueobject.photo_processing.photo_spot import PhotoSpot
 
 
 class ExifReader:
