+++ soil_analysis/domain/valueobject/photo_processing/photo_land_association.py
