+++ soil_analysis/domain/valueobject/photo_processing/photo_spot.py
