+++ soil_analysis/tests/domain/valueobject/photo/test_photo.py
@@ -2,9 +2,9 @@ import os
 from unittest import TestCase
 
 from lib.geo.valueobject.coord import XarvioCoord
-from soil_analysis.domain.valueobject.photo import AndroidPhoto
-from soil_analysis.domain.valueobject.photo import IphonePhoto
-from soil_analysis.domain.valueobject.photo_spot import PhotoSpot
+from soil_analysis.domain.valueobject.photo_processing.photo import AndroidPhoto
+from soil_analysis.domain.valueobject.photo_processing.photo import IphonePhoto
+from soil_analysis.domain.valueobject.photo_processing.photo_spot import PhotoSpot
 
 
 class TestBasePhotoFunctionality(TestCase):
